% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_analysis.R
\name{addMICAoutput}
\alias{addMICAoutput}
\title{Add the MICA output (cluster labels and UMAP/tSNE coordinates) to sparse eset object}
\usage{
addMICAoutput(input_eset, mica_output_file, visual_method = "umap")
}
\arguments{
\item{input_eset}{The sparse eset object to add the MICA output into}

\item{mica_output_file}{The .txt file generated by MICA. It includes 4 columns: "ID" (Cell ID), "X" (UMAP_1 or tSNE_1), "Y" (UMAP_2 or tSNE_2), "label" (ClusterID)}

\item{visual_method}{Character, method used for visualizing the clustering results: \code{umap} (the default) or \code{tsne}.}
}
\value{
A sparse eset object with clustering results added.
}
\description{
This function is used to add the clustering results by MICA into the sparse eset object. Two types of resluts would be added to the phenoData slot of sparse eset object:
\itemize{
\item cluster ID of each cell
\item dimension reduction coordinates of each cell, either \code{UMAP} ot \code{tSNE}.
}
}
\examples{
clustered.eset <- addMICAoutput(input_eset = input_eset, mica_output_file = "/path-to-mica-input/clustering_UMAP_euclidean_20_2.22554.txt", visual_method = "umap")
}
