% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_analysis.R
\name{cal_Activity}
\alias{cal_Activity}
\title{Calculate driver activities from gene expression matrix and networks}
\usage{
cal_Activity(
  target_list = NULL,
  cal_mat = NULL,
  activity_method = "mean",
  do.std = TRUE
)
}
\arguments{
\item{target_list}{a list of source genes returned by \code{get_net2target_list()}, for each source gene, the value if a data frame with "\code{target}", "\code{MI}" and "\code{spearman}" as the columns.}

\item{cal_mat}{a matrix of normalized and log-transformed gene expression values}

\item{activity_method}{Character, method used to calculate the activity: "\code{mean}" (the default), "\code{weightedmean}", "\code{absmean}" or "\code{maxmean}".}

\item{do.std}{Logical, whether to do the z-normalization on the gene expression values in each sample. Default: \code{TRUE}.}
}
\value{
a matrix of activities, drivers by cells
}
\description{
This function is used to calculate activities from gene expression matrix and networks. This is an internal function for \code{cal_Activity()}, \code{getActivity_individual()}, \code{getActivity_inBatch()}.
}
\examples{
act_mat <- cal_Activity(target_list = target_list, cal_mat = exprs(normalized.eset), activity_method = 'mean', do.std = TRUE)
}
