% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_analysis.R
\name{combinePvalVector}
\alias{combinePvalVector}
\title{Combine P values using Fisher's method or Stouffer's method}
\usage{
combinePvalVector(pvals, method = "Stouffer", signed = TRUE, twosided = TRUE)
}
\arguments{
\item{pvals}{A vector of numeric, P values to be combined or converted to Z scores}

\item{method}{Character, method used to combine P values: "\code{Stouffer}" (the default) or "\code{Fisher}".}

\item{signed}{Logical, whether the input P values are signed or not. Usually they are signed by folder change. Default: \code{TRUE}.}

\item{twosided}{Logical, whether the input P values are two-sided. If FALSE, the input P values will be treated as one-tailed. Default: \code{TRUE}.}
}
\value{
A vector containing "Z-statistics" and "P.Value".
}
\description{
This function is used to calculate the combined P value and Z score from multiple P values. it can also be used to convert P value to Z score.
}
\examples{
## 1. combine P values from a vector
combinePvalVector(c(0.1,1e-3,1e-5), method = 'Stouffer')

\dontrun{
## 2. combine P values from a table
df_Pcombined <- sapply(df$Pval*sign(df$log2FC), function(x) {combinePvalVector(x, twosided = TRUE)[2]})
df_Zscore <- sapply(df$Pval*sign(df$log2FC), function(x) {combinePvalVector(x, twosided = TRUE)[1]})
}

}
