% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.r
\name{draw.bubblePlot2}
\alias{draw.bubblePlot2}
\title{Inner function for simple bubbleplots}
\usage{
draw.bubblePlot2(df = NULL, xlab, ylab, clab, slab,
  low.col = "#004C99", high.col = "#CC0000", plot.title = NULL)
}
\arguments{
\item{df}{re-structured data.frame for bubble plots}

\item{xlab}{string}

\item{ylab}{string}

\item{clab}{string}

\item{slab}{string}

\item{low.col}{string,default as "#004C99"}

\item{high.col}{string, default as "CC0000}

\item{plot.title}{string}
}
\value{
a ggplot object
}
\description{
Inner function for simple bubbleplots
}
