% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{draw.marker.bbp}
\alias{draw.marker.bbp}
\title{Generate visualization for marker scores via bubble plot}
\usage{
draw.marker.bbp(ref = NULL, input_eset, feature = "geneSymbol",
  group_tag = "ClusterRes", save_plot = FALSE, width = 8,
  height = 5, plot_name = "AnnotationBubbleplot.png")
}
\arguments{
\item{ref}{reference dataframe, includes positive or negative markers for different cell types}

\item{input_eset}{expressionSet/SparseExpressionSet object with clustering membership stored in pData}

\item{feature}{feature type from your reference, should be in colnames(fData(eset))}

\item{group_tag}{a character, the variable containing clustering label in pData(eset)}

\item{save_plot}{logical, whether or not save your plot}

\item{width}{default as 8, inch as unit}

\item{height}{default as 5, inch as unit}

\item{plot_name}{plot name, please include plot type}
}
\value{
A ggplot object
}
\description{
Marker visualizatoin from known markers/signatures, requires knowledge-based marker list as input
}
\details{
draw.marker.bbp
}
