% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.r
\name{draw.scRNAseq.QC}
\alias{draw.scRNAseq.QC}
\title{draw.scRNAseq.QC}
\usage{
draw.scRNAseq.QC(SparseEset, project.name, plot.dir = "./QC/",
  output.cutoff = TRUE, group = "group")
}
\arguments{
\item{SparseEset}{an SparseEset generated by CreateSparseEset}

\item{project.name}{a character, project name to print on report}

\item{plot.dir}{a character, output directory for QC reports}
}
\value{
an R markdown QC report
}
\description{
generated a scRNA-seq quality control report in html with Rmarkdown
}
\details{
draw.scRNAseq.QC
}
