% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{draw_barplot}
\alias{draw_barplot}
\title{Bar plot showing the cell composition of self-defined groups}
\usage{
draw_barplot(
  input_eset,
  group_by = "clusterID",
  color_by = "cell_type",
  colors = NULL,
  legend.position = "right",
  xlabel.angle = 0,
  fontsize.legend_title = 12,
  fontsize.legend_text = 10,
  fontsize.axis_title = 12,
  fontsize.axis_text = 10
)
}
\arguments{
\item{input_eset}{The expression set object that filtered, normalized and log-transformed}

\item{group_by}{Character, name of the column for grouping, usually the column of cell types or clusters. Default: "\code{clusterID}".}

\item{color_by}{Character, name of the column for color-coding, usually the column of cell types or clusters. Default: "\code{cell_type}".}

\item{colors}{A vector of colors for filling the violins. The length should be same as the number of groups. Default: \code{NULL} (ggplot default colors).}

\item{legend.position}{Character, position of legend: "\code{right}" (the default), "\code{left}", "\code{top}", "\code{bottom}" or "\code{none}".}

\item{xlabel.angle}{Numeric, the angle of the a-axis title. When it's set not 0, the x-axis text will automatically right-justified. Default: 0.}

\item{fontsize.legend_title}{Integer, font size of the legend title. Default: 10.}

\item{fontsize.legend_text}{Integer, font size of the legend text. Default: 8.}

\item{fontsize.axis_title}{Integer, font size of the axis label and text. Default: 10.}

\item{fontsize.axis_text}{Integer, font size of the axis label and text. Default: 8.}
}
\value{
A ggplot object that can be visualized by "\code{p}" or \code{ggsave(file = "output.pdf", p)}
}
\description{
This function is used to draw a bar plot showing the cell composition of self-defined groups.
}
\examples{
data(pbmc14k_expression.eset)

## 1. bar plot grouped by clusters ("clusterID") and colored by true labels ("true_label)
draw_barplot(input_eset = pbmc14k_expression.eset,
             group_by = "clusterID",
             color_by = "true_label")

## 2. customize the colors
draw_barplot(input_eset = pbmc14k_expression.eset,
             group_by = "clusterID",
             color_by = "true_label",
             colors = c("green", "red", "blue", "grey", "orange", "purple", "yellow"))
}
