% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{draw_bubbleplot}
\alias{draw_bubbleplot}
\title{Bubble plot showing the signature scores by self-defined groups}
\usage{
draw_bubbleplot(
  input_eset,
  signature_table = NULL,
  group_by = "clusterID",
  colors = NULL,
  legend.position = "right",
  fontsize.legend_title = 10,
  fontsize.legend_text = 8,
  fontsize.axis_title = 10,
  fontsize.axis_text = 8,
  xlabel.angle = 0
)
}
\arguments{
\item{input_eset}{The expression set object that filtered, normalized and log-transformed}

\item{signature_table}{A matrix or data frame containing three columns: signature_name, signature_feature, weight. Default: \code{NULL}.}

\item{group_by}{Character, name of the column for grouping, usually the column of cell types or clusters. Default: "\code{clusterID}".}

\item{colors}{A vector of two colors indicating the low and high values respectively. Default: c("\code{lightgrey}", "\code{red}").}

\item{legend.position}{Character, position of legend: "\code{right}" (the default), "\code{left}", "\code{top}", "\code{bottom}" or "\code{none}".}

\item{fontsize.legend_title}{Integer, font size of the legend title. Default: 10.}

\item{fontsize.legend_text}{Integer, font size of the legend text. Default: 8.}

\item{fontsize.axis_title}{Integer, font size of the axis label and text. Default: 10.}

\item{fontsize.axis_text}{Integer, font size of the axis label and text. Default: 8.}

\item{xlabel.angle}{Numeric, the angle of the a-axis title. When it's set not 0, the x-axis text will automatically right-justified. Default: 0.}
}
\value{
A ggplot object of bubble plot
}
\description{
This function is used to draw a bubble plot of signature scores among self-defined groups.
}
\examples{
marker_file <- system.file('PBMC14KDS_DemoDataSet/DATA/', 'Immune_signatures.xlsx', package = "scMINER")
signature_table <- openxlsx::read.xlsx(marker_file)
head(signature_table)
## 1. the most commonly used command
draw_bubbleplot(input_eset = clustered.eset,
                signature_table = signature_table,
                group_by = "clusterID")

## 2. customize the colors
draw_bubbleplot(input_eset = clustered.eset,
                signature_table = signature_table,
                group_by = "clusterID",
                colors = c("lightgrey", "red"))
}
