% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{feature_boxplot}
\alias{feature_boxplot}
\title{Box plot showing the expression or activity of selected features by self-defined groups}
\usage{
feature_boxplot(
  input_eset,
  features = NULL,
  group_by = "clusterID",
  ncol = 3,
  colors = NULL,
  legend.position = "right",
  fontsize.legend_title = 10,
  fontsize.legend_text = 8,
  fontsize.strip = 10,
  fontsize.axis_title = 10,
  fontsize.axis_text = 8,
  xlabel.angle = 0,
  ylabel.text = "Expression (log2CPM)",
  stat_method = "median",
  add_jitter = FALSE,
  jitter.height = 0,
  jitter.width = 0.3,
  jitter.size = 0.1
)
}
\arguments{
\item{input_eset}{The expression set object that filtered, normalized and log-transformed}

\item{features}{A vector of genes or drivers (row.names of the input eset) to plot}

\item{group_by}{Character, name of the column for grouping, usually the column of cell types or clusters. Default: "\code{clusterID}".}

\item{ncol}{Integer, number of columns when multiple plots are displayed. Default: 3.}

\item{colors}{A vector of colors for filling the violins. The length should be same as the number of groups. Default: \code{NULL} (ggplot default colors).}

\item{legend.position}{Character, position of legend: "\code{right}" (the default), "\code{left}", "\code{top}", "\code{bottom}" or "\code{none}".}

\item{fontsize.legend_title}{Integer, font size of the legend title. Default: 10.}

\item{fontsize.legend_text}{Integer, font size of the legend text. Default: 8.}

\item{fontsize.strip}{Integer, font size of the plot strip. Default: 10.}

\item{fontsize.axis_title}{Integer, font size of the axis label and text. Default: 10.}

\item{fontsize.axis_text}{Integer, font size of the axis label and text. Default: 8.}

\item{xlabel.angle}{Numeric, the angle of the a-axis title. When it's set not 0, the x-axis text will automatically right-justified. Default: 0.}

\item{ylabel.text}{Character, the title of y-axis. Default: "\code{Expression (log2CPM)}"}

\item{stat_method}{Character or NULL. method of the stat point to show: "\code{median}" (the default), "\code{mean}". If \code{NULL}, the stat point won't show up.}

\item{add_jitter}{Logical, whether to add jittered points. Default: \code{FALSE}.}

\item{jitter.height}{Numeric, amount of vertical jitter. Default: 0.}

\item{jitter.width}{Numeric, amount of horizontal jitter. Default: 0.3.}

\item{jitter.size}{Numeric, size of the jittered points. Default: 0.1.}
}
\value{
A ggplot object with one or multiple box plots
}
\description{
This function is used to draw a box plot of selected features among self-defined groups from a sparse eset object.
}
\examples{
## 1. violin plots grouped by clusters (say the column name is 'clusterID')
p_box <- feature_boxplot(input_eset = clustered.eset, features = c("CD14", "CD19", "CD8A"), group_by = "clusterID")

## 2. violin plots grouped by cell types (say the column name is 'cellType')
p_box <- feature_boxplot(input_eset = clustered.eset, features = c("CD14", "CD19", "CD8A"), group_by = "cellType")

## 3. customize the colors to fill the violin plots
p_box <- feature_boxplot(input_eset = clustered.eset, features = c("CD14", "CD19", "CD8A"), group_by = "clusterID", colors = c("blue", "red", "green"))

## 4. add jittered points
p_box <- feature_boxplot(input_eset = clustered.eset, features = c("CD14", "CD19", "CD8A"), group_by = "clusterID", add_jitter = TRUE, jitter.width = 0.5, jitter.size = 0.5)

## 5. using activity data
p_box <- feature_boxplot(input_eset = activity_clustered.eset, features = c("CD14_SIG", "CD19_SIG", "CD8A_SIG"), group_by = "clusterID", ylabel_text = "Activity")
}
