% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{feature_bubbleplot}
\alias{feature_bubbleplot}
\title{Bubble blot showing the expression or activity of selected features by self-defined groups}
\usage{
feature_bubbleplot(
  input_eset,
  features = NULL,
  group_by = "clusterID",
  colors = NULL,
  legend.position = "right",
  fontsize.legend_title = 10,
  fontsize.legend_text = 8,
  fontsize.axis_title = 10,
  fontsize.axis_text = 8,
  xlabel.angel = 0
)
}
\arguments{
\item{input_eset}{The expression set object that filtered, normalized and log-transformed}

\item{features}{A vector of genes or drivers (row.names of the input eset) to plot}

\item{group_by}{Character, name of the column for grouping, usually the column of cell types or clusters. Default: "\code{clusterID}".}

\item{colors}{A vector of two colors indicating the low and high values respectively. Default: c("\code{lightgrey}", "\code{red}").}

\item{legend.position}{Character, position of legend: "\code{right}" (the default), "\code{left}", "\code{top}", "\code{bottom}" or "\code{none}".}

\item{fontsize.legend_title}{Integer, font size of the legend title. Default: 10.}

\item{fontsize.legend_text}{Integer, font size of the legend text. Default: 8.}

\item{fontsize.axis_title}{Integer, font size of the axis label and text. Default: 10.}

\item{fontsize.axis_text}{Integer, font size of the axis label and text. Default: 8.}

\item{xlabel.angel}{Numeric, the angel of the a-axis title. When it's set not 0, the x-axis text will automatically right-justified. Default: 0.}
}
\value{
A ggplot object of bubble plot
}
\description{
This function is used to draw a bubble plot of selected features among self-defined groups from a sparse eset object.
}
\examples{
features_of_interest <- c("CD3D","CD27","IL7R","SELL","CCR7","IL32","GZMA","GZMK","DUSP2","CD8A","GZMH","GZMB","CD79A","CD79B","CD86","CD14")
## 1. the most commonly used command
p_bubble <- feature_bubbleplot(input_eset = clustered.eset, features = features_of_interest, group_by = "clusterID")

## 2. customize the colors
p_bubble <- feature_bubbleplot(input_eset = clustered.eset, features = features_of_interest, group_by = "clusterID", colors = c("lightgrey", "red"))
}
