% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.r
\name{feature_heatmap}
\alias{feature_heatmap}
\title{Visualize gene expression level on scRNA-seq data}
\usage{
feature_heatmap(eset, target, feature = "geneSymbol",
  group_tag = "label", name = "log2Exp", save_plot = TRUE,
  width = 4, height = 8, cluster_rows = FALSE,
  plot_name = "GeneHeatmap.png", ...)
}
\arguments{
\item{eset}{Input expression set}

\item{feature}{character, which feature to visualize}

\item{group_tag}{a character, label to visualize on the top of heatmap}

\item{name}{character, name of value visualized in color scale}

\item{save_plot}{logical, whether to save plots or not}

\item{width}{numerical}

\item{height}{numerical}

\item{cluster_rows}{logical, if or not cluster rows}

\item{plot_name}{character, name of heamap}
}
\description{
This plot will visualiz feature info in scatter plot by outputing a ggplot object
}
