% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.r
\name{feature_highlighting}
\alias{feature_highlighting}
\title{Visualize gene expression level on scRNA-seq data}
\usage{
feature_highlighting(input_eset = eset, target, feature = "geneSymbol",
  x = "tSNE_1", y = "tSNE_2", ylabel = "Expression",
  title.size = 5)
}
\arguments{
\item{input_eset}{Input expression set}

\item{target}{a character vector, the list of feature to visualize}

\item{feature}{character, which feature to visualize}

\item{x}{cordinates for x axis}

\item{y}{cordinates for y axis}

\item{ylabel}{a characterm, title of y axis}

\item{title.size}{numerical, default as 5}
}
\description{
This plot will visualiz feature info in scatter plot by outputing a ggplot object
}
