% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{feature_scatterplot}
\alias{feature_scatterplot}
\title{Scatter plot showing the expression or activity of selected features on UMAP or t-SNE coordinates}
\usage{
feature_scatterplot(
  input_eset,
  features = NULL,
  location_x = "UMAP_1",
  location_y = "UMAP_2",
  colors = NULL,
  ncol = 3,
  point.size = 0.5,
  legend.position = "right",
  fontsize.legend_title = 10,
  fontsize.legend_text = 8,
  fontsize.strip = 10,
  fontsize.axis_title = 10,
  fontsize.axis_text = 8
)
}
\arguments{
\item{input_eset}{The expression set object that filtered, normalized and log-transformed}

\item{features}{A vector of genes or drivers (row.names of the input eset) to plot}

\item{location_x}{Character, name of the column of x-axis coordinates. Default: "\code{UMAP_1}".}

\item{location_y}{Character, name of the column of y-axis coordinates. Default: "\code{UMAP_2}".}

\item{colors}{A vector of two colors indicating the low and high values respectively. Default: c("\code{lightgrey}", "\code{red}").}

\item{ncol}{Integer, number of columns when multiple plots are displayed. Default: 3.}

\item{point.size}{Numeric, size of the scatter points. Default: 0.5.}

\item{legend.position}{Character, position of legend: "\code{right}" (the default), "\code{left}", "\code{top}", "\code{bottom}" or "\code{none}".}

\item{fontsize.legend_title}{Integer, font size of the legend title. Default: 10.}

\item{fontsize.legend_text}{Integer, font size of the legend text. Default: 8.}

\item{fontsize.strip}{Integer, font size of the plot strip. Default: 10.}

\item{fontsize.axis_title}{Integer, font size of the axis label and text. Default: 10.}

\item{fontsize.axis_text}{Integer, font size of the axis label and text. Default: 8.}
}
\value{
Print a plot to screen and return a gtable containing a list of plots, can be visualized by plot(g), and saved by ggsave(file = "output.pdf", g)
}
\description{
This function is used to draw a scatter plot of selected features on UMAP or t-SNE coordinates from a sparse eset object.
}
\examples{
## 1. scatter plots with UMAP projections
p_scatter <- feature_scatterplot(input_eset = clustered.eset, features = c("CD14", "CD19", "CD8A"), location_x = "UMAP_1", location_y = "UMAP_2")

## 2. scatter plots with t-SNE projections
p_scatter <- feature_scatterplot(input_eset = clustered.eset, features = c("CD14", "CD19", "CD8A"), location_x = "tSNE_1", location_y = "tSNE_2")

## 3. change the point size and font size
p_scatter <- feature_scatterplot(input_eset = clustered.eset, features = c("CD14", "CD19", "CD8A"), location_x = "UMAP_1", location_y = "UMAP_2", point.size = 1, fontsize.strip = 12, fontsize.axis = 10)
}
