% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMINER.R
\name{feature_vlnplot}
\alias{feature_vlnplot}
\title{feature_vlnplot}
\usage{
feature_vlnplot(
  input_eset,
  target = NULL,
  feature = "geneSymbol",
  group_by = "celltype",
  ylabel = "Expression",
  color_by = NULL,
  colors = NULL,
  ncol = 3,
  stat = "median",
  boxplot = FALSE,
  title.size = 5
)
}
\arguments{
\item{input_eset}{Input expression set}

\item{target}{a character vector, the list of feature to visualize}

\item{feature}{character, which feature to visualize}

\item{group_by}{character, which group info to visualize as x axis}

\item{ylabel}{a character, title of y axis}

\item{color_by}{character, which group info to define color, if NULL, then violin plots will be colored by 'group_by'}

\item{colors}{character vector, default as NULL, will use ggplot default color palette}

\item{ncol}{cordinates for y axis}

\item{stat}{a character, whether to plot median or mean as a black dot on violinplot}

\item{boxplot}{logical, whether to plot boxplot on violinplot}

\item{title.size}{numerical, default as 5}
}
\description{
This plot will visualize feature info in violin plot by outputing a ggplot object
}
\examples{
demo_file <- system.file('PBMC14KDS_DemoDataSet/DATA/pbmc.14k.DS.eset.log2.RData',
                        package = "scMINER")
load(demo_file)
genes_of_interest <-c("CD3D", "CD27", "IL7R",
                      "SELL", "CCR7", "IL32",
                      "GZMA", "GZMK", "DUSP2",
                      "CD8A", "GZMH", "GZMB",
                      "CD79A", "CD79B", "CD86", "CD14")
scMINER::feature_vlnplot(input_eset = pbmc.14k.DS.eset.log2,
                         target = genes_of_interest,
                         feature = "geneSymbol",
                         group_by = "ClusterRes",
                         ylabel = "log2Exp", ncol = 4)
}
