% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{generateMICAcmd}
\alias{generateMICAcmd}
\title{Generate MICA command script for job submission on local or LSF}
\usage{
generateMICAcmd(save_sh_at, input_file, project_name = "test",
  num_cluster = c(3, 4, 5), output_path, host = "lsf",
  queue = "standard", config_file = NULL, bootstrap = 10,
  dim_reduction_method = "MDS", visualization = "tsne",
  perplexity = 30, min_dist = 0.01, slice_size = 1000)
}
\arguments{
\item{save_sh_at}{path to save your MICA command file;}

\item{input_file}{path to actual input file for MICA, usually use \code{generateMICAinput} to generete file with desired format.}

\item{project_name}{character, name of your project, will be used for naming of output data}

\item{num_cluster}{a vector or a numerical number, the number of clusters}

\item{output_path}{path to MICA output file}

\item{host}{character, whether you want to run MICA pipeline on "lsf" or "local"}

\item{queue}{character. If host="lsf", which queue to submit your run_MICA.sh, default as NULL}

\item{config_file}{path to your customized config file where you could define queue, requeseted memory and processor; if use default file, then keep as NULL; defualt as NULL.}

\item{bootstrap}{number of iterations of k-means process, default as 10.}

\item{dim_reduction_method}{character, default as "mds". Other supported methods include "pca" and "lpl".}

\item{visualization}{character, visualization method used for visualize final clustering results. default as "tsne", other supported methods includes "umap"}

\item{perplexity}{numeical. parameter used in tsne visualization to indicate plot density}

\item{min_dist}{numerical. parameter used in umap visualization to indicate the distance between neighbours.}

\item{slice_size}{numerical. MICA sliced original data matrix into smaller pieces to improve computational efficiency,this parameter was used to indicate how many cells/samples should be grouped together as one downsized sub-matrix.
Default as 1000.}
}
\value{
A .sh script with runnable command for MICA execution
}
\description{
Generate command for running MICA locally or on LSF.
MICA is a high-performance clustering analysis method that was implemented in python and cwl.
}
\examples{
\dontrun{
generateMICAcmd<-function(  save_sh_at="./", #path to save shell script
                            input_file="MICA_input.txt", #your MICA input file
                            project_name="test",
                            num_cluster=c(3,4,5), #a vector of numerical number
                            output_path="./",
                            host="local", #or local
                            dim_reduction_method="MDS",
                            visualization="tsne")
}
}
