% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_analysis.R
\name{generateMICAinput}
\alias{generateMICAinput}
\title{Generate the standard input files for MICA from sparse eset object}
\usage{
generateMICAinput(
  input_eset,
  output_file,
  overwrite = F,
  downSample_N = NULL,
  seed = 1
)
}
\arguments{
\item{input_eset}{The sparse eset object to generate MICA input from. It must be normalized and log-transformed.}

\item{output_file}{The output file, or MICA input file. Should be in either "\strong{.txt}" or "\strong{.h5ad}" format.}

\item{overwrite}{Logical, whether to overwrite the output_file if it already exists. Default: \code{FALSE}.}

\item{downSample_N}{A non-negative integer or \code{NULL}, number of cells to downsample to. Default: \code{NULL}.}

\item{seed}{Integer or \code{NULL}, the seed for sampling. Default: 1. Ignored if \code{downSample_N} = \code{NULL}.}
}
\value{
A .txt or .h5ad file that can be used as the MICA input
}
\description{
This function is used to generate the standard input files for MICA (Mutual Information-based Clustering Analysis) from a sparse eset object. It
supports two file formats, "\strong{.txt}" or "\strong{.h5ad}". To generate a "\strong{.h5ad}" file, the "\strong{anndata}" package is required.
}
\examples{
data(pbmc14k_expression.eset)
generateMICAinput(input_eset = pbmc14k_expression.eset,
                  output_file = "/work-path/PBMC14k/MICA/micaInput.txt",
                  overwrite = FALSE)
}
