% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMINER.R
\name{generateSJARACNeInput}
\alias{generateSJARACNeInput}
\title{Generate SJARACNE input with designed folder structure}
\usage{
generateSJARACNeInput(
  input_eset,
  ref = NULL,
  funcType = NULL,
  input_driver = NULL,
  sampleN = 1000,
  seed = 1,
  wd.src,
  group_name,
  symbolColumnName = "geneSymbol"
)
}
\arguments{
\item{input_eset}{An expressionSet}

\item{ref}{c("hg", "mm"), could be a manually defined geneSymbol vector}

\item{funcType}{c("TF","SIG", NULL), if NULL then both TF and SIG will be considered}

\item{input_driver, }{a list of drivers for calculation. If NULL, curated driver list in the R package will be used.}

\item{sampleN}{integer, number of cells that should be sampled per group for computational efficiency purpose, default is 1000. If sampleN=NULL, no sampling will be performed.}

\item{seed}{integer, the random seed for sampling, default is 1.}

\item{wd.src}{output path}

\item{group_name}{name of group for sample identification}

\item{symbolColumnName}{name for the column that save gene symbols.}
}
\value{
SJARACNe input files for each subgroups
}
\description{
This function helps to generate appropriate input files for SJARACNe pipeline.
It can take transcription factor/signaling gene reference from internal(stored in package) or external (manual define)
}
\details{
generateSJARACNeInput
}
\examples{
demo_file <- system.file('PBMC14KDS_DemoDataSet/DATA/pbmc.14k.DS.eset.log2.RData',
                        package = "scMINER")
load(demo_file)
SJAR.cmd.tf <- generateSJARACNeInput(
               input_eset = pbmc.14k.DS.eset.log2,
               funcType = "TF",
               ref = "hg",  # human
               wd.src = 'test/',  # output directory
               group_name = "celltype")
\dontrun{
}
}
\keyword{SJARACNe}
