% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MINIE.R
\name{get.DA}
\alias{get.DA}
\title{Find differential activity genes from activity matrix}
\usage{
get.DA(
  input_eset = NULL,
  group_name = "celltype",
  group_case = NULL,
  group_ctrl = NULL,
  method = "t.test"
)
}
\arguments{
\item{input_eset}{ExpressionSet that stores group information in Biobase::pData}

\item{group_name}{a character string, column name in Biobase::pData(input_eset) that indicates group info}

\item{group_case}{NULL(If do get.DA for all group vs others) or a character string (one specific group vs others) of
column name in Biobase::pData(input_eset) that indicates group info}

\item{group_ctrl}{NULL(If one vs Others); a character indicate case group if do pairwise analysis}

\item{method}{a character from c("t.test", "limma"), which method will be used to identify differential activity gene}
}
\value{
output would be a data.frame containing: t.statistics, p.value, log2FC, z.score, and mean Activity value
}
\description{
\code{get.DA} is a wraper of (\code{DAG_test}, and \code{getDE.limma}),
which helps to conduct two_sided t.test on all genes in specific group VS Others
to find differential activity genes, a table with essential statistics will be outputted.
}
\examples{
demo_file <- system.file('PBMC14KDS_DemoDataSet/DATA/celltype_Activity.RData',
                                             package = "scMINER")
load(demo_file)
DAG_result_tf <- get.DA(input_eset = AC_eset$AC.TF,
                       group_name = "celltype")
}
