% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get.Topdrivers}
\alias{get.Topdrivers}
\title{get.Topdrivers}
\usage{
get.Topdrivers(DAG_result = DAG_result, n = 5, degree_filter = c(50,
  500), celltype = NULL)
}
\arguments{
\item{DAG_result}{Output table from function FindDAG}

\item{n}{threshold to pick top master regulators(top n)}

\item{degree_filter}{filter out drivers with target number less than certain value}

\item{celltype}{character, output top hits are from which celltype}
}
\value{
A list of top master regulators among different groups
}
\description{
Help quick pick top master regulators from previous
differential activity analysis results
}
