% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MINIE.R
\name{get_activity}
\alias{get_activity}
\title{Calculate activity from network file or gene list}
\usage{
get_activity(
  Net = NULL,
  eset,
  tag = NULL,
  genelist = NULL,
  use.symbol = FALSE,
  feature = "geneSymbol",
  es.method = "mean",
  activity.method = "weighted",
  normalize = TRUE,
  sep.symbol = "."
)
}
\arguments{
\item{Net}{Network data frame}

\item{eset}{ExpressionSet/SparseExpressionSet with expression data}

\item{tag}{If network is TF network or SIG network}

\item{genelist}{A list of signature gene list}

\item{use.symbol}{logical, in network file, use geneSymbol or use geneID}

\item{feature}{character, use which feature as ID in Biobase::fData(eset)}

\item{es.method}{character, which method to use to calculate actiivty value ("mean","maxmean")}

\item{activity.method}{character, which method to use to estimate activity ("weighted","unweighted")}

\item{normalize}{logical, if normalize or not}

\item{sep.symbol}{which symbol to sparate name and tag}
}
\value{

}
\description{
Calculate activity from network file or gene list
}
\details{
If network object was loaded by get.network.scMINER function, then network dataframe is could be retrieved under network_dat slot.
}
\examples{
demo_file <- system.file('PBMC14KDS_DemoDataSet/DATA/pbmc.14k.DS.eset.log2.RData',
                        package = "scMINER")
load(demo_file)
TF_file <- system.file('PBMC14KDS_DemoDataSet/SJAR/Bcell_11546_11539_77/tf_final/consensus_network_ncol_.txt',
            package = "scMINER")
tf.network  <- get.network.scMINER(network_file=TF_file)
acs1<-get_activity(Net = tf.network$network_dat,tag = "TF",
         normalize=T,
         eset = pbmc.14k.DS.eset.log2,
         activity.method = 'unweighted',
         use.symbol=TRUE)
}
