% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_analysis.R
\name{get_target_list2matrix}
\alias{get_target_list2matrix}
\title{Convert ther list generated by \code{get_net2target_list()} to a matrix of signed mutual information}
\usage{
get_target_list2matrix(target_list = NULL, activity_method = "mean")
}
\arguments{
\item{target_list}{a list of source genes returned by \code{get_net2target_list()}, for each source gene, the value if a data frame with "\code{target}", "\code{MI}" and "\code{spearman}" as the columns,}

\item{activity_method}{Character, method used to calculate the activity: "\code{mean}" (the default), "\code{weightedmean}", "\code{absmean}" or "\code{maxmean}".}
}
\value{
If "\code{weightedmean}" is given, it returns a matrix of signed (by Spearman correlation coefficient) mutual information, the sign of which will be used for activity calculation. For
all the other methods ("\code{mean}", "\code{absmean}" or "\code{maxmean}"), it returns a matrix of 1.
}
\description{
This function is used to conver the list generated by \code{get_net2target_list()} to a matrix of mutual information signed by Spearman correlation coefficient, sources by targets. This is an
internal function for c\code{al_Activity()}, \code{getActivity_individual()}, \code{getActivity_inBatch()}.
}
\examples{
net_data <-  read.table(system.file("extdata/demo_pbmc14k/SJARACNe/B/TF/bt100_pc001/consensus_network_ncol_.txt", package = "scMINER"),
                        header = TRUE, sep = "\t", stringsAsFactors = FALSE, quote = "",
                        colClasses = c("character", "character", "character", "character", "numeric", "numeric", "numeric", "numeric", "numeric"))
target_list <- get_net2target_list(net_data)

weight_matrix <- get_target_list2matrix(target_list)

}
