% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_analysis.R
\name{getActivity_inBatch}
\alias{getActivity_inBatch}
\title{Calculate driver activities in batch from the SJARACNe directory}
\usage{
getActivity_inBatch(
  input_eset,
  sjaracne_dir,
  group_name,
  group_exclude = NULL,
  network_tag.tf = NULL,
  network_tag.sig = NULL,
  driver_type = "TF_SIG",
  activity_method = "mean",
  do.z_normalization = TRUE
)
}
\arguments{
\item{input_eset}{The expression set object which has been filtered, normalized and log-transformed}

\item{sjaracne_dir}{The path to the SJARACNe directory}

\item{group_name}{Character, name of the column for grouping, usually the column of cell types or clusters}

\item{group_exclude}{A vector of group names to exclude in activity calculation}

\item{network_tag.tf}{Character or \code{NULL}, the tag used to distinguish different SJARACNe runs from the same input files. This is usually the name of the folder(s) in "\code{TF}" folder of each group. Default: \code{NULL}.}

\item{network_tag.sig}{Character or \code{NULL}, the tag used to distinguish different SJARACNe runs from the same input files. This is usually the name of the folder(s) in "\code{SIG}" folder of each group. Default: \code{NULL}.}

\item{driver_type}{Character, type of the pre-defined driver list to use. Should be one of: "\code{TF}" for transcriptional factors only, "\code{SIG}" for signaling genes only, or "\code{TF_SIG}" for both. Default: "\code{TF_SIG}".}

\item{activity_method}{Character, method used to calculate the activity: "\code{mean}" (the default), "\code{weightedmean}", "\code{absmean}" or "\code{maxmean}".}

\item{do.z_normalization}{Logical, whether to do the z-normalization on the gene expression values in each sample. Set if to \code{FALSE} only when the expression values has been scaled in each cell. Default: \code{TRUE}.}
}
\value{
A expression set object. The \code{assayData} is the activity matrix of all cells of all groups, drivers by cells. The \code{phenoData} and \code{featureData} are exactly save with the input eset.
}
\description{
This function is used to calculate the driver activities of multiple groups from a scMINER directory. To calculate driver activities of one single group, please use \code{getActivity_individual()}.
}
\examples{
data(pbmc14k_expression.eset)

## 1. when no tag was used in runing SJARACNE: the network file folder ("sjaracne_workflow-*") is directly under TF/SIG folder of each group.
activity.eset <- getActivity_inBatch(input_eset = pbmc14k_expression.eset,
                                     sjaracne_dir = "./SJARACNe",
                                     group_name = "cell_type",
                                     driver_type = "TF_SIG",
                                     activity_method = "mean",
                                     do.z_normalization = TRUE)

## 2. when tag (e.g. "bs_100" ) was used: the nework file folder ("sjaracne_workflow-*") is directly under a subfolder "bs_100" of the TF/SIG folder of each group.
activity.eset <- getActivity_inBatch(input_eset = pbmc14k_expression.eset,
                                     sjaracne_dir = "./SJARACNe",
                                     group_name = "cell_type",
                                     network_tag.tf = "bs_100",
                                     network_tag.sig = "bs_100",
                                     driver_type = "TF_SIG",
                                     activity_method = "mean",
                                     do.z_normalization = TRUE)

## 3. to calculate the activities of TF only
activity.eset <- getActivity_inBatch(input_eset = pbmc14k_expression.eset,
                                     sjaracne_dir = "./SJARACNe",
                                     group_name = "cell_type",
                                     network_tag.tf = "bs_100",
                                     network_tag.sig = "bs_100",
                                     driver_type = "TF",
                                     activity_method = "mean",
                                     do.z_normalization = TRUE)

## 4. to exclude some groups in the activity calculation (e.g. "NK" and "Monocyte")
activity.eset <- getActivity_inBatch(input_eset = pbmc14k_expression.eset,
                                     sjaracne_dir = "./SJARACNe",
                                     group_name = "cell_type",
                                     group_exclude = c("NK", "Monocyte"),
                                     network_tag.tf = "bs_100",
                                     network_tag.sig = "bs_100",
                                     driver_type = "TF",
                                     activity_method = "mean",
                                     do.z_normalization = TRUE)

## 5. when calculate the activities from the gene expression values scaled by other methods (e.g. ScaleData() from Seurat package)
activity.eset <- getActivity_inBatch(input_eset = pbmc14k_expression.eset,
                                     sjaracne_dir = "./SJARACNe",
                                     group_name = "cell_type",
                                     network_tag.tf = "bs_100",
                                     network_tag.sig = "bs_100",
                                     driver_type = "TF_SIG",
                                     activity_method = "mean",
                                     do.z_normalization = FALSE)
}
