% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MINIE.R
\name{GetActivityFromSJARACNe}
\alias{GetActivityFromSJARACNe}
\title{GetActivityFromSJARACNe}
\usage{
GetActivityFromSJARACNe(
  SJARACNe_output_path = NA,
  SJARACNe_input_eset = NA,
  functype = "tf",
  group_name = NA,
  activity.method = "unweighted",
  activity.norm = TRUE,
  save_network_file = FALSE,
  save_path = NULL
)
}
\arguments{
\item{SJARACNe_output_path}{Path to SJARACNe output folder(s)}

\item{SJARACNe_input_eset}{Expressionset that you generate input from}

\item{functype}{character c("tf","sig"); If NULL, both activity from TF and SIG network will be calculated;
default as NULL}

\item{group_name}{a string, group name stored in Biobase::pData that defines expression matrix separation}

\item{activity.method}{c("weighted,unweighted), default to "unweighted"}

\item{activity.norm}{logical, default to TRUE.}

\item{save_network_file}{logical, default to FALSE}

\item{save_path}{Path to save network file}
}
\value{
An expressionset with activity values
}
\description{
Allocate network information from SJARACNe and calculate activity score for each hub genes.
}
\examples{
demo_file <- system.file('PBMC14KDS_DemoDataSet/DATA/pbmc.14k.DS.eset.log2.RData',
                        package = "scMINER")
load(demo_file)
out.dir.SJAR <- system.file('PBMC14KDS_DemoDataSet/SJAR/',
                        package = "scMINER")
acs.14k.tf <- GetActivityFromSJARACNe(
              SJARACNe_output_path = out.dir.SJAR,
              SJARACNe_input_eset = pbmc.14k.DS.eset.log2,
              activity.method="unweighted",
              activity.norm=TRUE,
              group_name =  "celltype",
              save_network_file=FALSE,
              functype="tf",
              save_path='.')
\dontrun{
}

}
\author{
Chenxi Qian, \email{chenxi.qian@stjude.org}
}
\keyword{GetActivity}
