% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_analysis.R
\name{getDA}
\alias{getDA}
\title{Perform differential activity analysis on expression set}
\usage{
getDA(
  input_eset,
  group_by = "clusterID",
  g1 = NULL,
  g0 = NULL,
  use_method = "t.test"
)
}
\arguments{
\item{input_eset}{The expression set object that filtered, normalized and log-transformed}

\item{group_by}{Character, name of the column for grouping, usually the column of cell types or clusters. Default: "\code{clusterID}".}

\item{g1}{A vector of character defining the fore-ground group or \code{NULL}. Default: \code{NULL}.}

\item{g0}{A vector of character defining the back-ground group or \code{NULL}. Default: \code{NULL}.}

\item{use_method}{Character, method used for differential analysis: "\code{limma}", "\code{wilcoxon}", and "\code{t.test}" (the default).}
}
\value{
A data frame. Rows are genes/drivers, and columns are 11 statistics of differential analysis.
}
\description{
Perform differential activity analysis on expression set
}
\examples{
## 1. To perform differential activity analysis in a 1-vs-rest manner for all groups in "clusterID" column
da_res <- getDA(input_eset = activity_clustered.eset,
                group_by = "clusterID",
                use_method = "t.test")

## 2. To perform differential activity analysis in a 1-vs-rest manner for one specific group in "clusterID" column
da_res <- getDA(input_eset = activity_clustered.eset,
                group_by = "clusterID",
                g1 = c("1"),
                use_method = "t.test")

## 3. To perform differential activity analysis in a rest-vs-1 manner for one specific group in "clusterID" column
da_res <- getDA(input_eset = activity_clustered.eset,
                group_by = "clusterID",
                g0 = c("1"),
                use_method = "t.test")

## 4. To perform differential activity analysis in a 1-vs-1 manner for groups in "clusterID" column
da_res <- getDA(input_eset = activity_clustered.eset,
                group_by = "clusterID",
                g1 = c("1"),
                g0 = c("3"),
                use_method = "t.test")
}
