% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_analysis.R
\name{getDriverList}
\alias{getDriverList}
\title{Extract the pre-defined driver lists of human or mouse}
\usage{
getDriverList(species_type = "hg", driver_type = "TF")
}
\arguments{
\item{species_type}{Character, species of the driver lists to be extracted: "\code{hg}" for human, "\code{mm}" for mouse. Default: "\code{hg}".}

\item{driver_type}{Character, type of drivers to be extracted: "\code{TF}" for transcriptional factor, "\code{SIG}" for signaling genes, and "\code{TF_SIG}" for both. Default: "\code{TF}".}
}
\value{
A vector of pre-defined driver genes
}
\description{
This function is used to extract the pre-defined driver lists of human or mouse.
}
\examples{
hg_tf <- getDriverList(species_type = "hg", driver_type = "TF") # get the TF driver list of human
mm_driver <- getDriverList(species_type = "mm", driver_type = "TF_SIG") # get the total driver list, including both TF and SIG, of mouse
}
