% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_analysis.R
\name{getTopFeatures}
\alias{getTopFeatures}
\title{Pick the top genes/drivers for each group from differential analysis results}
\usage{
getTopFeatures(
  input_table,
  number = 10,
  group_by = "g1_tag",
  sort_by = "log2FC",
  sort_decreasing = TRUE
)
}
\arguments{
\item{input_table}{The table generated by \code{getDE()} or \code{getDA()}, containing 11 statistics of differential analysis.}

\item{number}{An Integer, number of top genes/driver to pick from each group. Default: 10.}

\item{group_by}{Character, name of the column for grouping, usually the column of cell types or clusters. Default: "\code{g1_tag}".}

\item{sort_by}{Character, name of the column for sorting. Default: "\code{log2FC}".}

\item{sort_decreasing}{Logical, whether to sort the column specified by \code{sort_by} in decreasing order. If \code{FALSE}, the column will be sorted in increasing order. Default: \code{TRUE}.}
}
\value{
A data frame with top genes/driver of each group
}
\description{
This function is used to pick the top genes/drivers for each group from differential analysis results based on either \verb{fold change}, \verb{p value} or \code{FDR}.
}
\examples{
\dontrun{
top_drivers <- getTopFeature(da_res, number = 10, group_by, "g1_tag")
}

}
