% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.r
\name{MICAplot}
\alias{MICAplot}
\title{plot MICA}
\usage{
MICAplot(input_eset, label = NULL, visualize = NULL, X = NULL,
  Y = NULL, title.size = 10, title.name = "", pct = 0.5)
}
\arguments{
\item{input_eset}{ExpressionSet that include visualization coordinates in phenotype data}

\item{label}{Coloring criteria of data points, should be a variable stored in pData(input_eset)}

\item{visualize}{character, name of visualization method, this will be used on x or y axis title}

\item{X}{character, column name of x axis}

\item{Y}{character, column name of y axis}

\item{title.size}{numerical, size of plot title, default as 10}

\item{title.name}{character, title of plot, default as NULL}

\item{pct}{numerical, size of point, default as 0.5}
}
\description{
This function helps to generate a ggplot object for phenotypic visualization
}
