% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_sparseEset.R
\name{normalizeSparseEset}
\alias{normalizeSparseEset}
\title{Normalize and log-transform the sparse eset object}
\usage{
normalizeSparseEset(
  input_eset,
  scale_factor = 1e+06,
  do.logTransform = TRUE,
  log_base = 2,
  log_pseudoCount = 1
)
}
\arguments{
\item{input_eset}{The sparse eset object for normalization}

\item{scale_factor}{Numeric, the library size to normalize to. Default: 1000000.}

\item{do.logTransform}{Logical, whether to do log-transformation. Default: \code{TRUE}.}

\item{log_base}{Numeric, the base of log-transformation. Usually \strong{2}, \strong{exp(1)} or \strong{10}. Default: 2.}

\item{log_pseudoCount}{Numeric, the pseudo count to add to avoid "\code{-Inf}" in log-transformation. Default: 1.}
}
\value{
A sparse eset object that has been normalized and log-transformed
}
\description{
This function is used to normalize and log-transform the sparse eset object. The default method is "log21p".
}
\examples{
data("pbmc14k_rawCount")
pbmc14k_raw.eset <- createSparseEset(input_matrix = pbmc14k_rawCount, projectID = "PBMC14k", addMetaData = TRUE)
pbmc14k_filtered.eset <- filterSparseEset(pbmc14k_raw.eset, filter_mode = "auto", filter_type = "both")

pbmc14k_log2cpm.eset <- normalizeSparseEset(pbmc14k_filtered.eset,
                                            scale_factor = 1000000,
                                            log_base = 2,
                                            log_pseudoCount = 1)

}
