% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc14k_expression.eset}
\alias{pbmc14k_expression.eset}
\title{SparseEset object of PBMC14k dataset}
\format{
\subsection{\code{pbmc14k_expression.eset}}{

A large dgCMatrix with 17,986 rows and 14,000 columns:
\describe{
This data set provides the SparseEset object of PBMC14k dataset that has been filtered, normalized, clustered and annotated.
}
}
}
\source{
It's generated by scMINER from Filtered_DownSampled_SortedPBMC_data.csv from \url{https://zenodo.org/record/3357167#.YhQNF2RKj6V}
}
\usage{
pbmc14k_expression.eset
}
\description{
This dataset contains the SparseEset object of PBMC14k dataset. For demonstration purposes, it has been downsampled to 3.5k cells, with 500 cells per population.
}
\keyword{datasets}
