% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc14k_MICAoutput}
\alias{pbmc14k_MICAoutput}
\title{MICA output for PBMC14k dataset}
\format{
\subsection{\code{pbmc14k_MICAoutput}}{

A data frame with 13,605 rows and 4 columns:
\describe{
\item{ID}{barcodes of cells}
\item{X}{coordinates of x-axis, usually UMAP_1 or tSNE_1}
\item{Y}{coordinates of y-axis, usually UMAP_2 or tSNE_2}
\item{label}{labels of predicted clusters by MICA}
}
}
}
\source{
generated by MICA
}
\usage{
pbmc14k_MICAoutput
}
\description{
This is an example of clustering label file for PBMC14k dataset generated by MICA.
}
\keyword{datasets}
