% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc14k_signatureTable}
\alias{pbmc14k_signatureTable}
\title{Signature table of 7 known cell types of PBMC14k dataset}
\format{
\subsection{\code{pbmc14k_signatureTable}}{

A data frame with 31 rows and 3 columns:
\describe{
\item{signature_name}{name of signatures/cell types}
\item{signature_feature}{genes/features of corresponding signatures}
\item{weight}{weight of each feature for corresponding signature, ranging from -1 to 1}
}
}
}
\source{
Self-curated from multiple sources
}
\usage{
pbmc14k_signatureTable
}
\description{
This dataset contains the self-curated markers of 7 known cell types of PBMC14k dataset
}
\keyword{datasets}
