% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc14k_trueLabel}
\alias{pbmc14k_trueLabel}
\title{True label of known cell types for PBMC14k dataset}
\format{
\subsection{\code{pbmc14k_trueLabel}}{

A data frame with 14,000 rows and 2 columns:
\describe{
\item{cell_barcode}{barcodes of cells}
\item{true_label}{true label cell types of cells}
}
}
}
\source{
A subset of Labels.csv from \url{https://zenodo.org/record/3357167#.YhQNF2RKj6V}
}
\usage{
pbmc14k_trueLabel
}
\description{
This dataset contains the true label of known cell types of 14,000 Peripheral Blood Mononuclear Cells (PBMCs) in PBMC14k dataset.
}
\keyword{datasets}
