% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMINER.R
\name{preMICA.filtering}
\alias{preMICA.filtering}
\title{preMICA.filtering}
\usage{
preMICA.filtering(
  SparseEset,
  cutoffs,
  gene_filter = TRUE,
  nGene_filter = TRUE,
  nUMI_filter = TRUE,
  ERCC_filter = TRUE,
  Mito_filter = TRUE
)
}
\arguments{
\item{SparseEset}{the sparseEset object outputted from draw.scRNAseq.QC}

\item{cutoffs}{a list outputted by draw.scRNAseq.QC, if NULL, manual input will be required}

\item{gene_filter}{logical; or a numerical number, indicating lower threshold for gene filtering based on how many non-zero cells each gene expressed in}

\item{nGene_filter}{logical; a numerical number, indicating lower threshold put on number of gene expression in each cell for cell filtering}

\item{nUMI_filter}{logical;a vector of two numerical number, indicating lower threshold and upper threshold put on number of total UMI for cell filtering}

\item{ERCC_filter}{logical; a numerical number, indicating upper threshold put on ERCC percentage for cell filtering}

\item{Mito_filter}{logical;a numerical number, indicating upper threshold put on Mitochondrial gene expression fraction for cell filtering}
}
\value{
A Sparse expression set
}
\description{
scRNA-seq filtering function
}
\details{
preMICA.filtering
}
\examples{
demo_file <- system.file('PBMC14KDS_DemoDataSet/DATA/pbmc.14k.DS.eset.log2.RData',
                        package = "scMINER")
load(demo_file)
cutoffs <- draw.scRNAseq.QC(SparseEset = pbmc.14k.DS.eset.log2,
                                     project.name = project_name,
                                     plot.dir = 'test/',
                                     group = "group",
                                     output.cutoff = TRUE,
                                     only.cutoff = TRUE)
pbmc.14k.DS.eset.filter <- preMICA.filtering(SparseEset = pbmc.14k.DS.eset.log2,
      cutoffs = cutoffs)
}
