% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMINER.R
\name{readInput_10x.dir}
\alias{readInput_10x.dir}
\title{Read 10X Genomics data from a directory}
\usage{
readInput_10x.dir(
  input_dir,
  featureType = "gene_symbol",
  removeSuffix = TRUE,
  addPrefix = NULL
)
}
\arguments{
\item{featureType}{Character string, specify either gene_symbol or gene_id to use as the gene name. Default: gene_symbol.}

\item{removeSuffix}{Logical scalar, whether to remove the suffix "-1" when present in all cell barcodes. Default: TRUE.}

\item{addPrefix}{Character string or NULL, add a prefix to the cell barcodes, like sample ID. It is highly recommended to use a prefix containing letters and/or numbers only, and not starting with numbers.}

\item{dir}{Directory containing the 3 files generated by 10x Genomics: matrix.mtx, barcodes.tsv and features.tsv (or genes.tsv).}
}
\value{
A sparse gene expression matrix of raw UMI counts, gene by cell
}
\description{
Read 10X Genomics data from a directory
}
\examples{
input_dir <- 'path-to-directory'
list.files(input_dir, full.names = FALSE)
sparseMatrix <- readInput_10x.dir(input_dir, featureType = "gene_symbol", removeSuffix = TRUE, addPrefix = "demoSample")
}
