% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_input.R
\name{readInput_10x.h5}
\alias{readInput_10x.h5}
\title{Read the input data generated by 10x Genomics from the HDF5 file}
\usage{
readInput_10x.h5(
  h5_file,
  featureType = "gene_symbol",
  removeSuffix = TRUE,
  addPrefix = NULL
)
}
\arguments{
\item{h5_file}{H5 file generated by CellRanger pipeline of 10x Genomics}

\item{featureType}{Character, feature type to use as the gene name of
expression matrix: \code{"gene_symbol"} (the default) or \code{"gene_id"}.}

\item{removeSuffix}{Logical, whether to remove the suffix "\strong{-1}" when
present in all cell barcodes. Default: \code{TRUE}.}

\item{addPrefix}{Character or \code{NULL}, add a prefix to the cell barcodes, like
Sample ID. It is highly recommended to use a prefix containing letters
and/or numbers only, and not starting with numbers. Default: \code{NULL}.}
}
\value{
A sparse gene expression matrix of raw UMI counts, genes by cells
}
\description{
This function is used to read the gene expression data from the
HDF5 file generated by CellRanger pipeline of 10x Genomics. This function can
automatically distinguish the data of different modalities (e.g. expression
data, ATAC data) and retains the gene expression data only. The \verb{**hdf5r**}
package is needed to use this function.
}
\examples{
h5_file <- system.file("extdata/demo_inputs/hdf5_10x/demoData3.h5", package = "scMINER") # path to hdf5 file
sparseMatrix <- readInput_10x.h5(h5_file,
                                 featureType = "gene_symbol",
                                 removeSuffix = TRUE,
                                 addPrefix = "demoSample")

}
