% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{readMICAoutput}
\alias{readMICAoutput}
\title{readMICAoutput}
\usage{
readMICAoutput(eset = NULL, input_file, output_file, load_ClusterRes = TRUE)
}
\arguments{
\item{eset}{a SparseMatrix Eset}

\item{input_file}{input expression txt file of MICA pipeline}

\item{output_file}{output ClusterMem.txt file from MICA pipeline}

\item{load_ClusterRes}{logical, if TRUE, clustering results will be store at pData(eset)$label}
}
\value{
A sparse expressionSet object
}
\description{
Read MICA input and output to create an expressionSet for downstream analysis
}
