% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{readscRNAseqData}
\alias{readscRNAseqData}
\title{readscRNAseqData}
\usage{
readscRNAseqData(file, is.10x = TRUE, CreateSparseEset = TRUE,
  add.meta = F, ...)
}
\arguments{
\item{file}{data path to 10x genomics output folder, which normally contains 3 files (matrix.mtx, gene or feature.tsv and barcode.csv),
or data path to data txt/csv/tsv file}

\item{is.10x}{logical, whether or not inputs are from CellRanger standard output}

\item{CreateSparseEset}{logical, whether or not create sparse matrix incorporated expression set}

\item{add.meta}{logical, whether or not calculate metadata info from expression matrixm, this is not suggested before merging/downsampling your data}

\item{...}{parameters pass to read.delim if is.10x = FALSE}
}
\value{
A list or sparse matrix expression set
}
\description{
read scRNA-seq data, a wrapper of conventional data reading (read.delim2) and 10x genomics data standarad output reading
}
