% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMINER.R
\name{scMINER.dir.create}
\alias{scMINER.dir.create}
\title{Manipulation of Working Directories for scMINER pipeline}
\usage{
scMINER.dir.create(project_main_dir = NULL, project_name = NULL)
}
\arguments{
\item{project_main_dir}{character, name or absolute path of the main working directory.}

\item{project_name}{character, name of the project folder.}
}
\value{
\code{scMINER.dir.create} returns a list object, containing main.dir (path of the main working directory),
project.name (project name), out.dir (path of the project folder).
}
\description{
\code{scMINER.dir.create} is used to help users create an organized working directory for the network construction step in scMINER analysis.
However, it is not essential for the analysis.
It creates a hierarchical working directory and returns a list contains this directory information.
}
\details{
This function needs users to define the main working directory and the project's name.
It creates a main working directory with a subdirectory of the project.
It also automatically creates five subfolders (DATA, SJAR, MICA, QC, PLOT) within the project folder.
DATA/, storing data files;
SJAR/, storing files needed for running SJAracne;
MICA/, storing files needed for running MICA;
QC/,   storing Quality Control related plots;
PLOT/, storing plot files;
This function also returns a list object (example, \code{scminer.par} in the demo) with directory information wrapped inside.
}
\examples{

\dontrun{
# Creating a main working directory under the current working directory by folder name
scminer.par <- scMINER.dir.create("MyMainDir","MyProject")
# Or creating a main working directory under the current working directory by relative path
scminer.par <- scMINER.dir.create("./MyMainDir","MyProject")
# Or creating a main working directory to a specific path by absolute path
scminer.par <- scMINER.dir.create("~/Desktop/MyMainDir","MyProject")
}
}
