% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_sparseEset.R
\name{updateSparseEset}
\alias{updateSparseEset}
\title{Update the slots and/or meta data of the sparse eset object}
\usage{
updateSparseEset(
  input_eset,
  dataMatrix = NULL,
  cellData = NULL,
  featureData = NULL,
  addMetaData = FALSE
)
}
\arguments{
\item{input_eset}{The sparse eset object to update}

\item{dataMatrix}{A data matrix with Features/Genes as the rows and Cells as the columns. It's row.names and colnames must be consistent with the \strong{\code{input_eset}}. Default: \code{NULL}.}

\item{cellData}{A data frame containing meta data of cells or \code{NULL}. It's row.names should be consistent with the colnames of \strong{\code{input_eset}}. Default: \code{NULL}.}

\item{featureData}{A data frame containing meata data of features or \code{NULL}. It's row.names should be consistent with the row.names of \strong{\code{input_eset}}. Default: \code{NULL}.}

\item{addMetaData}{Logical, whether to update the meta data of features and cells based on the expression matrix. Default: \code{FALSE}.}
}
\value{
A sparse eset object with updated information
}
\description{
This function is used to update the three slots ('\strong{assayData}', '\strong{phenoData}', '\strong{featureData}') and/or '\strong{meta data}' of sparse eset object.
}
\examples{
data("pbmc14k_expression.eset")

## 1. Update the QC metrics: this will recalculate the 'nUMI', 'nFeature', 'pctMito', 'pctSpikeIn' and 'nCell'. This is very helpful when the SparseEset is subsetted or combined.
pbmc14k_raw.eset <- updateSparseEset(input_eset = pbmc14k_expression.eset, addMetaData = TRUE)

## 2. Update the meta data of cells: with this, the user can add more information of cells, like clustering results, cell type.
true_label <- read.table(system.file("extdata/demo_pbmc14k/PBMC14k_trueLabel.txt.gz", package = "scMINER"), header = T, row.names = 1, sep = "\t", quote = "", stringsAsFactors = FALSE)
pbmc14k_raw.eset <- updateSparseEset(input_eset = pbmc14k_expression.eset, cellData = true_label, addMetaData = TRUE)

}
