% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_analysis.R
\name{z_normalization}
\alias{z_normalization}
\title{Scale a numeric vector using Z-normalization}
\usage{
z_normalization(x)
}
\arguments{
\item{x}{a vector of numbers for z-normalization}
}
\value{
a vector of numbers with z-normalized
}
\description{
This function is used to scale a numeric vector using Z-normalization. \code{NA} would be ignored. This is an internal function for \code{cal_Activity()}, \code{getActivity_individual()}, \code{getActivity_inBatch()}.
}
\examples{
z_normalization(c(1,2,3,4,5,NA,6,7,8,9))
}
