% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_analysis.R
\name{cal_Activity}
\alias{cal_Activity}
\title{Calculate driver activities from gene expression matrix and networks}
\usage{
cal_Activity(
  target_list = NULL,
  cal_mat = NULL,
  activity_method = "mean",
  do.std = TRUE
)
}
\arguments{
\item{target_list}{a list of source genes returned by \code{get_net2target_list()}, for each source gene, the value if a data frame with "\code{target}", "\code{MI}" and "\code{spearman}" as the columns.}

\item{cal_mat}{a matrix of normalized and log-transformed gene expression values}

\item{activity_method}{Character, method used to calculate the activity: "\code{mean}" (the default), "\code{weightedmean}", "\code{absmean}" or "\code{maxmean}".}

\item{do.std}{Logical, whether to do the z-normalization on the gene expression values in each sample. Default: \code{TRUE}.}
}
\value{
a matrix of activities, drivers by cells
}
\description{
This function is used to calculate activities from gene expression matrix and networks. This is an internal function for \code{cal_Activity()}, \code{getActivity_individual()}, \code{getActivity_inBatch()}.
}
\examples{
net_data <-  read.table(system.file("extdata/demo_pbmc14k/SJARACNe/B/TF/bt100_pc001/consensus_network_ncol_.txt", package = "scMINER"),
                        header = TRUE, sep = "\t", stringsAsFactors = FALSE, quote = "",
                        colClasses = c("character", "character", "character", "character", "numeric", "numeric", "numeric", "numeric", "numeric"))
target_list <- get_net2target_list(net_data)

data(pbmc14k_expression.eset)

act_mat <- cal_Activity(target_list = target_list, cal_mat = exprs(pbmc14k_expression.eset), activity_method = 'mean', do.std = TRUE)
}
