% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_sparseEset.R
\name{combineSparseEset}
\alias{combineSparseEset}
\title{Combine multiple sparse expression set objects}
\usage{
combineSparseEset(
  eset_list,
  projectID = NULL,
  addPrefix = NULL,
  addSurfix = NULL,
  addMetaData = TRUE,
  imputeNA = TRUE
)
}
\arguments{
\item{eset_list}{A vector of sparse expression set objects to combine}

\item{projectID}{A character vector or \code{NULL}, set the project names of the eset objects to combine. Default: \code{NULL}.}

\item{addPrefix}{A character vector or \code{NULL}, add a \strong{prefix} to the cell barcodes of each eset object to combine. It is highly recommended to use a prefix containing letters and/or numbers only, and not starting with numbers. Default: \code{NULL}.}

\item{addSurfix}{A character vector or \code{NULL}, add a \strong{surfix} to the cell barcodes of each eset object to combine. It is highly recommended to use a surfix containing letters and/or numbers only, and not starting with numbers. Default: \code{NULL}.}

\item{addMetaData}{Logical, whether to update the meta data of cells and features after combination. Default: \code{TRUE}.}

\item{imputeNA}{Logical, whether to impute NA values in combined matrix. If \code{TRUE}, the min value of the matrix will be used to replace the NAs. If \code{FALSE}, NA values will retain. Default: \code{TRUE}.}
}
\value{
A sparse eset object with combined features and cells of multiple eset objects.
}
\description{
This function is used to combine the sparse expression set objects. The combined eset object contains all cells and features of all input eset objects. If the eset objects are of
different features, \code{NA} values will be generated and automatically imputed by the minimum value of the combined gene expression matrix.
}
\examples{
combined.eset <- combineSparseEset(c(sample_1.eset, sample_2.eset, sample_3.eset), projectID = c("sample1", "sample2", "sample3"), addPrefix = c("tag1", "tag2", "tag3"), addMetaData = TRUE)
}
