% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_analysis.R
\name{compare2groups}
\alias{compare2groups}
\title{Perform differential analysis between two groups}
\usage{
compare2groups(
  input_eset,
  group_by = "clusterID",
  g1 = NULL,
  g0 = NULL,
  use_method = "limma"
)
}
\arguments{
\item{input_eset}{The expression set object that filtered, normalized and log-transformed}

\item{group_by}{Character, name of the column for grouping, usually the column of cell types or clusters. Default: "\code{clusterID}".}

\item{g1}{A vector of character defining the fore-ground group or \code{NULL}. Default: \code{NULL}.}

\item{g0}{A vector of character defining the back-ground group or \code{NULL}. Default: \code{NULL}.}

\item{use_method}{Character, method used for differential analysis: "\code{limma}" (the default), "\code{wilcoxon}", and "\code{t.test}".}
}
\value{
A data frame. Rows are genes/drivers, and columns are 11 statistics of differential analysis.
}
\description{
This function is used to perform the differential analysis between two groups. It supports there methods: "\code{limma}", "\code{wilcoxon}", and "\code{t.test}". This is an internal function for \code{getDE()} and \code{getDA()}.
}
\examples{
## to call this function
res <- compare2groups(input_eset = input_eset, group_by = group_by, g1 = g1_tmp, g0 = g0_tmp, use_method = use_method)
}
