% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MINIE.R
\name{ConvertNet2List}
\alias{ConvertNet2List}
\title{Convert Pairwise Network Data Frame to Driver-to-Target List}
\usage{
ConvertNet2List(net_dat = NULL)
}
\arguments{
\item{net_dat}{data.frame, must contain two columns with column names "source" (driver) and "target" (target genes).
"MI" (mutual information) and "spearman" (spearman correlation coefficient) columns are optional, but strongly suggested to use.
If "MI" and "spearman" columns are missing, errors may occur in some following steps (e.g. es.method='weightedmean' in \code{cal.Activity}).}
}
\value{
Return a list. The names of the list elements are drivers.
Each element is a data frame, contains three columns. "target", target gene names;
"MI", mutual information; "spearman", spearman correlation coefficient.
}
\description{
\code{ConvertNet2List} is a helper function in the \code{get.SJAracne.network}.
But if users have their own pairwise gene network files, they can convert it to driver-to-target list object.
}
\examples{
\dontrun{
tf.network.file <- sprintf('\%s/demo1/network/SJAR/project_2019-02-14/\%s/\%s',
                   system.file(package = "NetBID2"),
                   'output_tf_sjaracne_project_2019-02-14_out_.final',
                   'consensus_network_ncol_.txt')
net_dat      <- read.delim(file=tf.network.file,stringsAsFactors = FALSE)
target_list  <- ConvertNet2List(net_dat)
}
}
