% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{DAG_ttest}
\alias{DAG_ttest}
\title{DAG_ttest}
\usage{
DAG_ttest(d, group)
}
\arguments{
\item{d}{A vector of gene expression}

\item{group}{A vector of group information}
}
\description{
Inner function to do t.test(pairwise/2case) from activity matrix
}
