% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMINER.R
\name{draw.group.barplot}
\alias{draw.group.barplot}
\title{Draw barplot for composition study}
\usage{
draw.group.barplot(input_eset, group_by, color_by, colors = NULL)
}
\arguments{
\item{input_eset}{ExpressionSet that include group information in phenotype data}

\item{group_by}{Group criteria for bars, should be a variable stored in pData(input_eset)}

\item{color_by}{Coloring criteria of bar fractions, should be a variable stored in pData(input_eset)}

\item{colors}{color values to feed in scale_fill_manual, default as NULL; If NULL, then default color for ggplot will be used}
}
\value{
a ggplot object
}
\description{
Draw barplot for composition study
}
