% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMINER.R
\name{draw.marker.bbp}
\alias{draw.marker.bbp}
\title{Generate visualization for marker scores via bubble plot}
\usage{
draw.marker.bbp(
  ref = NULL,
  input_eset,
  feature = "geneSymbol",
  group_name = "ClusterRes",
  save_plot = FALSE,
  width = 8,
  height = 5,
  plot_name = "AnnotationBubbleplot.png"
)
}
\arguments{
\item{ref}{reference dataframe, includes positive or negative markers for different cell types;
Specify first column as different cell types, second columns as markers, third columns as weight (postive or negative marker)}

\item{input_eset}{expressionSet/SparseExpressionSet object with clustering membership stored in Biobase::pData}

\item{feature}{feature type from second column of your reference , should be in colnames(Biobase::fData(eset))}

\item{group_name}{a character, the variable containing clustering label in Biobase::pData(eset); or any other group information stored in Biobase::pData(eset)}

\item{save_plot}{logical, whether or not save your plot; if TRUE, plot will be saved as plot_name}

\item{width}{default as 8, inch as unit}

\item{height}{default as 5, inch as unit}

\item{plot_name}{plot name, please include plot type}
}
\value{
A ggplot object
}
\description{
Marker visualizatoin from known markers/signatures, requires knowledge-based marker list as input
}
\details{
Visualize marker score of different cell types on bubbleplot
}
\examples{
demo_file <- system.file('PBMC14KDS_DemoDataSet/DATA/pbmc.14k.DS.eset.log2.RData',
                        package = "scMINER")
load(demo_file)
markers_file <- system.file('PBMC14KDS_DemoDataSet/DATA/',
 'Immune_signatures.xlsx',
  package = "scMINER")
markers <- openxlsx::read.xlsx(markers_file)
draw.marker.bbp(ref = markers, input_eset = pbmc.14k.DS.eset.log2,
               width = 6, height = 4, feature = "geneSymbol",
               group_name = "ClusterRes", save_plot = FALSE)
\dontrun{
}

}
