% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMINER.R
\name{draw.scRNAseq.QC}
\alias{draw.scRNAseq.QC}
\title{draw.scRNAseq.QC}
\usage{
draw.scRNAseq.QC(
  SparseEset,
  project.name,
  plot.dir = "./QC/",
  output.cutoff = TRUE,
  group = "group",
  only.cutoff = FALSE
)
}
\arguments{
\item{SparseEset}{an SparseEset generated by CreateSparseEset}

\item{project.name}{a character, project name to print on report}

\item{plot.dir}{a character, output directory for QC reports}

\item{output.cutoff}{logical, whether or not return a list of suggested thresholds for filtering}

\item{group}{a character, a variable name indicate groupping information (stored in Biobase::pData) to
help generate violin plots}

\item{only.cutoff}{logical, whether or not only return cuttoff without plotting QC}
}
\value{
an R markdown QC report and a list of suggested threshold (if specify)
}
\description{
generated a scRNA-seq quality control report in html with Rmarkdown
}
\details{
draw.scRNAseq.QC
}
\examples{
demo_file <- system.file('PBMC14KDS_DemoDataSet/DATA/pbmc.14k.DS.eset.log2.RData',
                        package = "scMINER")
load(demo_file)
cutoffs <- scMINER::draw.scRNAseq.QC(SparseEset = pbmc.14k.DS.eset.log2,
               project.name = 'test',
               plot.dir = '.',
               group = "group",      # this indicate which meta data information will be use in x axis to group violin plots
               output.cutoff = TRUE) # whether or not to output suggested cutoffs
}
