% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMINER.R
\name{draw.scRNAseq.QC}
\alias{draw.scRNAseq.QC}
\title{draw.scRNAseq.QC}
\usage{
draw.scRNAseq.QC(
  SparseEset,
  project.name,
  plot.dir = "./QC/",
  output.cutoff = TRUE,
  group = "group"
)
}
\arguments{
\item{SparseEset}{an SparseEset generated by CreateSparseEset}

\item{project.name}{a character, project name to print on report}

\item{plot.dir}{a character, output directory for QC reports}

\item{output.cutoff}{logical, whether or not return a list of suggested thresholds for filtering}

\item{group}{a character, a variable name indicate groupping information (stored in pData) to
help generate violin plots}
}
\value{
an R markdown QC report and a list of suggested threshold (if specify)
}
\description{
generated a scRNA-seq quality control report in html with Rmarkdown
}
\details{
draw.scRNAseq.QC
}
