% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_analysis.R
\name{drawNetworkQC}
\alias{drawNetworkQC}
\title{Assess the quality of each network generated by SJARACNe}
\usage{
drawNetworkQC(
  network_file = NULL,
  sjaracne_dir = NULL,
  directed = TRUE,
  weighted = TRUE,
  generate_html = TRUE,
  outdir = NULL,
  prefix = NULL
)
}
\arguments{
\item{network_file}{The path to the network file (\strong{\code{consensus_network_ncol_.txt}}) for quality control}

\item{sjaracne_dir}{The path to the folder of SJARACNe runs. When this argument is given, scMINER will automatically retrieve all network files (\strong{\code{consensus_network_ncol_.txt}}) in this folder and generate quality control report for all of them.}

\item{directed}{Logical, whether the network is directed or not. All networks generated by SJARACNe are directed. Default: \code{TRUE}.}

\item{weighted}{Logical, whether the edges of the network is weighted or not. All networks generated by SJARACNe are weighted by mutual information ("\code{MI}" column). Default: \code{TRUE}.}

\item{generate_html}{Logical, whether to generate a html report. Default: \code{TRUE}.}

\item{outdir}{Character or \code{NULL}, the path to save the html report. If \code{NULL}, the html report will be saved in the same folder of the network file. Default: NULL. Ignored if \code{generate_html} = \code{FALSE}.}

\item{prefix}{Character or \code{NULL}, the character string to add in front of the html report file name. Default: \code{NULL}. Ignored if \code{generate_html} = \code{FALSE}.}
}
\value{
This function will print the statistics of several key quality metrics of network(s). If \code{generate_html} is set \code{TRUE}, it also generates a html file of quality control report and save it to the same folder of the network file (by default) or the folder specified by "\code{outdir}".
}
\description{
This function is used to assess the quality of networks generated by SJARACNe. It returns a summary table of key statistics of networks. The users can also generate the html quality control report by turning \code{generate_html} = \code{TRUE}.
}
\examples{
\dontrun{
## 1. Assess the quality of network from network files
drawNetworkQC(network_file = "/project_space/SJARACNE/B/SIG/b100/consensus_network_ncol_.txt", generate_html = TRUE,
              outdir = "/folder-to-save-report_file", prefix = "PBMC14")

## 2. Assess the quality of network from the directory of network files
drawNetworkQC(sjaracne_dir = "/project_space/SJARACNE", generate_html = TRUE)
}

}
