% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_sparseEset.R
\name{drawSparseEsetQC}
\alias{drawSparseEsetQC}
\title{Generate a quality control report from sparse eset object}
\usage{
drawSparseEsetQC(
  input_eset,
  output_html_file,
  overwrite = FALSE,
  group_by = NULL
)
}
\arguments{
\item{input_eset}{The sparse eset object for quality control analysis}

\item{output_html_file}{The path of the output .html file}

\item{overwrite}{Logical, whether to overwrite the output .html file if it already exists. Default: \code{FALSE}.}

\item{group_by}{Character or \code{NULL}, Name of column in pData(eset) used for grouping. Default: \code{NULL}.}
}
\value{
A html-formatted quality control report of sparse eset object
}
\description{
This function is used to generate a html quality control report from a sparse eset object. Compared with the summary table return by filterSparseEset(), the output report
contains more comprehensive and detailed QC results and can be used to estimate the cutoffs to filter the eset object. It also contains some plots for presentation purpose.
}
\examples{
## 1. To generate the QC report in a group-specific manner, recommended whenever group information is avaiable.
drawSparseEsetQC(input_eset = pbmc14k_raw.eset,
                 output_html_file = "/your-path/PBMC14k/PLOT/pbmc14k_rawCount.html",
                 overwrite = FALSE,
                 group_by = "trueLabel")

## 2. To generate the QC report from a whole view
drawSparseEsetQC(input_eset = pbmc14k_raw.eset,
                 output_html_file = "/your-path/PBMC14k/PLOT/pbmc14k_rawCount.html",
                 overwrite = FALSE,
                 group_by = NULL)
}
