% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMINER.R
\name{feature_heatmap}
\alias{feature_heatmap}
\title{Visualize gene expression level on scRNA-seq data via heatmap}
\usage{
feature_heatmap(
  input_eset,
  target,
  feature = "geneSymbol",
  group_name = "label",
  name = "log2Exp",
  save_plot = TRUE,
  width = 4,
  height = 8,
  cluster_rows = FALSE,
  colors = rev(colorRampPalette(brewer.pal(10, "RdYlBu"))(256)),
  plot_name = "GeneHeatmap.png",
  ...
)
}
\arguments{
\item{input_eset}{Input expression set}

\item{target}{a character or a character vector indicating feature names}

\item{feature}{a character, which feature to visualize}

\item{group_name}{a character, label to visualize on the top of heatmap}

\item{name}{character, name of value visualized in color scale}

\item{save_plot}{logical, whether to save plots or not}

\item{width}{numerical}

\item{height}{numerical}

\item{cluster_rows}{logical, if or not cluster rows}

\item{colors}{color palette}

\item{plot_name}{character, name of heamap}

\item{...}{parameter to be passed to ComplexHeatmap::Heatmap}
}
\value{
a ggplot object
}
\description{
This plot will visualiz feature info in scatter plot by outputing a ggplot object
}
\examples{
demo_file <- system.file('PBMC14KDS_DemoDataSet/DATA/pbmc.14k.DS.eset.log2.RData',
                        package = "scMINER")
load(demo_file)
genes_of_interest <-c("CD3D", "CD27", "IL7R",
                      "SELL", "CCR7", "IL32",
                      "GZMA", "GZMK", "DUSP2",
                      "CD8A", "GZMH", "GZMB",
                      "CD79A", "CD79B", "CD86", "CD14")
scMINER::feature_heatmap(input_eset = pbmc.14k.DS.eset.log2,
                     target = genes_of_interest,
                     group_name = "ClusterRes",
                     save_plot = FALSE,
                     width = 6, height = 6,
                     name = "log2Exp")
}
