% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMINER.R
\name{feature_highlighting}
\alias{feature_highlighting}
\title{feature_highlighting}
\usage{
feature_highlighting(
  input_eset,
  target = NULL,
  feature = "geneSymbol",
  x = "X",
  y = "Y",
  wrap_by = NULL,
  ylabel = "Expression",
  pct.size = 0.8,
  title.size = 15,
  ncol = 4,
  alpha = 0.8,
  colors = colorRampPalette(c("#E3E3E3", "#BCA2FC", "#4900FE"), interpolate =
    "linear")(8)
)
}
\arguments{
\item{input_eset}{Input expression set}

\item{target}{a character vector, the list of feature to visualize}

\item{feature}{character, which feature to visualize}

\item{x}{cordinates for x axis}

\item{y}{cordinates for y axis}

\item{wrap_by}{character, variable to wrap plot with}

\item{ylabel}{a characterm, title of y axis}

\item{pct.size}{numrical, point size}

\item{title.size}{numerical, default as 5}

\item{ncol}{cordinates for y axis}

\item{alpha}{numerical, default as 0.8}

\item{colors}{color palette for feature highlighting}
}
\description{
This plot will visualize feature info on scatter plot by outputing a ggplot object
}
\examples{
demo_file <- system.file('PBMC14KDS_DemoDataSet/DATA/pbmc.14k.DS.eset.log2.RData',
                        package = "scMINER")
load(demo_file)
genes_of_interest <-c("CD3D", "CD27", "IL7R",
                      "SELL", "CCR7", "IL32",
                      "GZMA", "GZMK", "DUSP2",
                      "CD8A", "GZMH", "GZMB",
                      "CD79A", "CD79B", "CD86", "CD14")
scMINER::feature_highlighting(input_eset = pbmc.14k.DS.eset.log2,
                      target = genes_of_interest,
                      feature = "geneSymbol",
                      ylabel = "log2Exp",
                      x = "X", y = "Y",
                      pct.size = 0.5)
}
