% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{feature_highlighting}
\alias{feature_highlighting}
\title{feature_highlighting}
\usage{
feature_highlighting(input_eset, target = NULL, feature = "geneSymbol",
  x = "X", y = "Y", visualize = "tSNE", ylabel = "Expression",
  pct.size = 0.8, title.size = 15, ncol = 4, alpha = 0.8,
  colors = colorRampPalette(c("#E3E3E3", "#BCA2FC", "#4900FE"),
  interpolate = "linear")(8))
}
\arguments{
\item{input_eset}{Input expression set}

\item{target}{a character vector, the list of feature to visualize}

\item{feature}{character, which feature to visualize}

\item{x}{cordinates for x axis}

\item{y}{cordinates for y axis}

\item{visualize}{character, name of visualization method, this will be used as x or y axis label}

\item{ylabel}{a characterm, title of y axis}

\item{pct.size}{numrical, point size}

\item{title.size}{numerical, default as 5}

\item{ncol}{cordinates for y axis}

\item{alpha}{numerical, default as 0.8}

\item{colors}{color palette for feature highlighting}
}
\description{
This plot will visualize feature info on scatter plot by outputing a ggplot object
}
