% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_sparseEset.R
\name{filterSparseEset}
\alias{filterSparseEset}
\title{Filter the cells and/or features of sparse eset object using automatic or self-customized cutoffs}
\usage{
filterSparseEset(
  input_eset,
  filter_mode = "auto",
  filter_type = "both",
  gene.nCell_min = NULL,
  gene.nCell_max = NULL,
  cell.nUMI_min = NULL,
  cell.nUMI_max = NULL,
  cell.nFeature_min = NULL,
  cell.nFeature_max = NULL,
  cell.pctMito_min = NULL,
  cell.pctMito_max = NULL,
  cell.pctSpikeIn_min = NULL,
  cell.pctSpikeIn_max = NULL
)
}
\arguments{
\item{input_eset}{The sparse eset object to be filtered}

\item{filter_mode}{Character, mode to apply the filtration cutoffs: \strong{"auto"} (the default) or \strong{"manual"}.}

\item{filter_type}{Character, objective type to be filtered: \strong{"both"} (the default), \strong{"cell"} or \strong{"feature"} .}

\item{gene.nCell_min}{Numeric, the minimum number of cells that the qualified genes are identified in. Default: 1.}

\item{gene.nCell_max}{Numeric, the maximum number of cells that the qualified genes are identified in. Default: Inf.}

\item{cell.nUMI_min}{Numeric, the minimum number of total UMI counts per cell that the qualified cells carry. Default: 1.}

\item{cell.nUMI_max}{Numeric, the maximum number of total UMI counts per cell that the qualified cells carry. Default: Inf.}

\item{cell.nFeature_min}{Numeric, the minimum number of non-zero Features per cell that the qualified cells carry. Default: 1.}

\item{cell.nFeature_max}{Numeric, the maximum number of non-zero Features per cell that the qualified cells carry. Default: Inf.}

\item{cell.pctMito_min}{Numeric, the minimum percentage of UMI counts of mitochondrial genes that the qualified cells carry. Default: 0.}

\item{cell.pctMito_max}{Numeric, the maximum percentage of UMI counts of mitochondrial genes that the qualified cells carry. Default: 1.}

\item{cell.pctSpikeIn_min}{Numeric, the minimum percentage of UMI counts of spike-in that the qualified cells carry. Default: 0.}

\item{cell.pctSpikeIn_max}{Numeric, the maximum percentage of UMI counts of spike-in that the qualified cells carry. Default: 1.}
}
\value{
A filtered sparse eset object. It also prints the summary of filtration to the screen.
}
\description{
This function is used to remove the cells and features of low quality. It provides two modes to define the cutoffs:
\itemize{
\item \strong{"auto"}: in this mode, scMINER will estimate the cutoffs based on Median ± 3*MAD (maximum absolute deviation). This mode works well for the matrix of raw UMI counts or TPM (Transcripts Per Million) values.
\item \strong{"manual"}: in this mode, the users can manually specify the cutoffs, both low and high, of all 5 metrics: \strong{nUMI}, \strong{nFeature}, \strong{pctMito}, \strong{pctSpikeIn} for cells, and \strong{nCell} for genes. No cells or
features would be removed under the default cutoffs of each metrics.
}
}
\examples{
filtered.eset <- filterSparseEset(raw.eset) ## filter the input eset using the cutoffs calculated by scMINER.
filtered.eset <- filterSparseEset(raw.eset, gene.nCell_min = 10, cell.nUMI_min = 500, cell.nFeature_min = 100, cell.nFeature_max = 5000, cell.pctMito_max = 0.15)
}
