% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{generateMICAinput}
\alias{generateMICAinput}
\title{generateMICAinput}
\usage{
generateMICAinput(d, filename)
}
\arguments{
\item{d}{matrix with colnames as cell/sample info, rownames as gene/feature info}

\item{filename}{filename of your MICA input file, supported format: txt}
}
\description{
A utility function that helps generate MICA input from a data matrix with rownames and colnames
}
