% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{generateSJARACNeInput}
\alias{generateSJARACNeInput}
\title{Generate SJARACNE input with designed folder structure}
\usage{
generateSJARACNeInput(input_eset, ref = NULL, funcType = NULL, wd.src,
  group_name)
}
\arguments{
\item{input_eset}{An expressionSet}

\item{ref}{c("hg", "mm"), could be a manually defined geneSymbol vector}

\item{funcType}{c("TF","SIG", NULL), if NULL then both TF and SIG will be considered}

\item{wd.src}{output path}

\item{group_name}{name of group for sample identification}
}
\value{
SJARACNe input files for each subgroups
}
\description{
This function helps to generate appropriate input files for SJARACNe pipeline.
It can take transcription factor/signaling gene reference from internal(stored in package) or external (manual define)
}
\details{
generateSJARACNeInput
}
\examples{
\dontrun{
generateSJARACNeInput(input_eset = eset ,ref = "hg",funcType="TF",
wd.src = "./",group_name = "celltype")
}
}
\keyword{SJARACNe}
