% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MINIE.R
\name{get.Topdrivers}
\alias{get.Topdrivers}
\title{get.Topdrivers}
\usage{
get.Topdrivers(
  DAG_result = DAG_result,
  n = 5,
  degree_filter = c(50, 500),
  celltype = NULL
)
}
\arguments{
\item{DAG_result}{Output table from function FindDAG}

\item{n}{threshold to pick top master regulators(top n)}

\item{degree_filter}{filter out drivers with target number less than certain value}

\item{celltype}{character, output top hits are from which celltype}
}
\value{
A list of top master regulators among different groups
}
\description{
Help quick pick top master regulators from previous
differential activity analysis results
}
\examples{
demo_file <- system.file('PBMC14KDS_DemoDataSet/DATA/celltype_Activity.RData',
                                             package = "scMINER")
load(demo_file)
DAG_result_tf <- get.DA(input_eset = AC_eset$AC.TF,
                       group_name = "celltype")
celltype <- levels(Biobase::pData(AC_eset$AC.TF)[,"celltype"])
TF_list <- get.Topdrivers(DAG_result = DAG_result_tf,
                          celltype = celltype,
                          n = 5, degree_filter = c(50, 600))
}
