% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_analysis.R
\name{MICAplot}
\alias{MICAplot}
\title{Draw a scatter plot showing the coordinates and cluster id of each cell}
\usage{
MICAplot(
  input_eset,
  color_by = "clusterID",
  colors = NULL,
  X = "UMAP_1",
  Y = "UMAP_2",
  point.size = 0.3,
  point.alpha = 1,
  name.plot_title = NULL,
  fontsize.plot_table = 20,
  show.cluster_label = TRUE,
  fontsize.cluster_label = 12,
  legend.position = "right",
  fontsize.legend_title = 10,
  fontsize.legend_text = 8,
  fontsize.axis_title = 10,
  fontsize.axis_text = 8
)
}
\arguments{
\item{input_eset}{The sparse eset object}

\item{color_by}{Character, name of the column of MICA cluster labels. Default: "\code{clusterID}".}

\item{colors}{A character vector or \code{NULL}, colors of the MICA cluster labels. The length of this vector should be same as the number of groups in color_by column. If \code{NULL}, the ggplot default colors will be use. Default: \code{NULL}.}

\item{X, Y}{Character, name of the columns of x-axis and y-axis coordinates. Default: "\code{UMAP_1}", and "\code{UMAP_2}".}

\item{point.size}{Numeric, size of points. Default: 0.3.}

\item{point.alpha}{Numeric, transparency of points, ranging from 0 (more transparent) to 1 (less transparent). Default: 1.}

\item{name.plot_title}{Character or NULL, title of the plot. Default: \code{NULL}.}

\item{fontsize.plot_table}{Numeric, font size of the title. Default: 20.}

\item{show.cluster_label}{Logical, whether to show labels on the plot. Default: \code{TRUE}.}

\item{fontsize.cluster_label}{Numeric, font size of the labels. Default: 12.}

\item{legend.position}{Character, position of legend: "\code{right}", "\code{left}", "\code{top}", "\code{bottom}" or "\code{none}". Default: "\code{right}".}

\item{fontsize.legend_title}{Integer, font size of the legend title. Default: 10.}

\item{fontsize.legend_text}{Integer, font size of the legend text. Default: 8.}

\item{fontsize.axis_title}{Integer, font size of the axis title. Default: 10.}

\item{fontsize.axis_text}{Integer, font size of the axis text. Default: 8.}
}
\value{
A UMAP or T-SNE plot. It also print the plot to screen.
}
\description{
This function is used to visualize the clustering results generated by MICA.
}
\examples{
p_umap <- MICAplot(input_eset = clustered.eset, color_by = "clusterID", X = "UMAP_1", Y = "UMAP_2", point.size = 0.1)
}
