% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{MICAplot}
\alias{MICAplot}
\title{plot MICA clustering results or other meta variables}
\usage{
MICAplot(input_eset, color_by = "ClusterRes", colors = NULL,
  X = NULL, Y = NULL, show_label = FALSE, title.size = 20,
  title.name = "", pct = 0.5, alpha = 1)
}
\arguments{
\item{input_eset}{ExpressionSet that include visualization coordinates in phenotype data}

\item{color_by}{Coloring criteria of data points, should be a variable stored in pData(input_eset)}

\item{colors}{character, color values, if NULL then use ggplot default color}

\item{X}{character, column name of x axis}

\item{Y}{character, column name of y axis}

\item{show_label}{logical, whether or not to show label on tSNE plot}

\item{title.size}{numerical, size of plot title, default as 10}

\item{title.name}{character, title of plot, default as NULL}

\item{pct}{numerical, size of point, default as 0.5}

\item{aplha}{numerical, indicate point transparency}
}
\description{
This function helps to generate a ggplot object for phenotypic visualization
}
