% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filtering.R
\name{preMICA.filtering}
\alias{preMICA.filtering}
\title{preMICA.filtering}
\usage{
preMICA.filtering(SparseEset, cutoffs, gene_filter = T,
  nGene_filter = T, nUMI_filter = T, ERCC_filter = T,
  Mito_filter = T)
}
\arguments{
\item{cutoffs}{a list outputted by draw.scRNAseq.QC, if NULL, manual input will be required}

\item{gene_filter}{logical; or a numerical number, indicating lower threshold for gene filtering based on how many non-zero cells each gene expressed in}

\item{nGene_filter}{logical; a numerical number, indicating lower threshold put on number of gene expression in each cell for cell filtering}

\item{nUMI_filter}{logical;a vector of two numerical number, indicating lower threshold and upper threshold put on number of total UMI for cell filtering}

\item{ERCC_filter}{logical; a numerical number, indicating upper threshold put on ERCC percentage for cell filtering}

\item{Mito_filter}{logical;a numerical number, indicating upper threshold put on Mitochondrial gene expression fraction for cell filtering}

\item{Obj}{the list object outputted by draw.scRNAseq.QC}
}
\value{
A Sparse expression set
}
\description{
scRNA-seq filtering function
}
\details{
preMICA.filtering
}
