% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_input.R
\name{readInput_10x.dir}
\alias{readInput_10x.dir}
\title{Read the input data generated by 10x Genomics from a directory}
\usage{
readInput_10x.dir(
  input_dir,
  featureType = "gene_symbol",
  removeSuffix = TRUE,
  addPrefix = NULL
)
}
\arguments{
\item{input_dir}{Path to the directory containing the 3 files generated by 10x Genomics: \strong{matrix.mtx}, \strong{barcodes.tsv} and \strong{features.tsv} (or \strong{genes.tsv})}

\item{featureType}{Character, feature type to use as the gene name of expression matrix: \code{"gene_symbol"} (the default) or \code{"gene_id"}.}

\item{removeSuffix}{Logical, whether to remove the suffix "\strong{-1}" when present in all cell barcodes. Default: \code{TRUE}.}

\item{addPrefix}{Character or \code{NULL}, add a prefix to the cell barcodes, like Sample ID. It is highly recommended to use a prefix containing letters and/or numbers only,
and not starting with numbers. Default: \code{NULL}.}
}
\value{
A sparse gene expression matrix of raw UMI counts, genes by cells
}
\description{
This function is used to read the gene expression data from a directory containing three files generated by 10x Genomics: \strong{matrix.mtx}, \strong{barcodes.tsv} and
\strong{features.tsv} (or \strong{genes.tsv}). This function can handle these conditions well:
\itemize{
\item Alternative file names for feature data: \strong{features.tsv} by CellRanger > 3.0, and \strong{genes.tsv} by CellRanger < 3.0;
\item One or more input files are compressed, usually in "\strong{.gz}" format;
\item Data with multiple modalities: like the single cell multiome data. In this case, it only retains the data of "Gene Expression".
}
}
\examples{
input_dir <- 'path-to-directory'
list.files(input_dir, full.names = FALSE) # you should see three files: matrix.mtx, barcodes.tsv and features.tsv (or genes.tsv)
sparseMatrix <- readInput_10x.dir(input_dir, featureType = "gene_symbol", removeSuffix = TRUE, addPrefix = "demoSample")
}
