% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{SJARACNe_filter}
\alias{SJARACNe_filter}
\title{SJARACNe_filter}
\usage{
SJARACNe_filter(eset.sel, tf.ref, sig.ref, wd.src, grp.tag)
}
\arguments{
\item{eset.sel}{ExpressionSet to generate SJaracne input}

\item{tf.ref}{A vector of reference transcription factors}

\item{sig.ref}{A vector of reference signaling genes}

\item{wd.src}{path to store SjAracne input}

\item{grp.tag}{name of group for identification}
}
\value{
A folder with picked master regulator and filtered gene expression matrix
}
\description{
This is the inner function to help generate SJARACNe input for scRNA-seq data,
all non-informative (zero genes) will be filtered in by this function
}
\details{
Non-expressed genes in subgroups are filtered.
tf.ref should be coordinate with featureNames(eset.sel).
}
